Close multilines comment in JSDoc example
Change the comments to use &#47; instead of / so that we can close the multiline comment in the example. 

Because the mulitline comment was not closed, the code-snippet is not compiling if you copy/paste it, complaining about a missing '*/' and the error is present as well in the StackBlitz snippet.